+++ Changelog.md
@@ -13,3 +13,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added September 5, 2022
 - Footer Content
 - Adjust navbar spacing
+
+### Added September 5, 2022
+- Landing Page Content